Creating a "current" folder for the most current "bleeding-edge" development.
For all cs-* projects, the "current" folder should be compatible.  For instance, these should always be compatible (the FULL URL is probably https://{projectName}.svn.sourceforge.net/svnroot/{projectName}):
   -- cs-webapplibs/trunk/current
   -- cs-content/trunk/current
   -- cs-phpxml/trunk/current

This should help avoid conflicts for older checkouts that are still trying to be bleeding edge; it is pretty easy for one project (e.g. cs-webapplibs) to get too far behind or too far ahead of other projects without even realizing it, due to how the URLs are setup in the svn:externals property...